Cleanup Travis (#76)
* Stop testing PHP 7.1

This version is EOL.

* Stop testing Symfony 4.0 and 4.1

These versions are EOL.

* Use `dist: trusty` to fix issue with RabbitMQ

Sometimes, RabbitMQ would not start with error:
```
Failed to start rabbitmq-server.service: Unit rabbitmq-server.service not found.
```

See https://travis-ci.community/t/is-rabbitmq-service-down/4362

* Stop testing specific Doctrine versions

The logic was flawed anyway. It would pin ORM and DBAL to the same version, but they have different version ranges. DBAL is now on 2.10 and ORM on 2.7.